fix: add condition if total file less or equal than keep file

+++ index.js
@@ -19,12 +19,16 @@ const del = async () => {
     try {
         const totalFile = await execCommand(`ls -1t ${dir}/* | wc -l`);
         const totalDeleteFiles = totalFile - keepFile;
+        if (totalFile <= keepFile) {
+            return false;
+        }
         const deletedFiles = await execCommand(`ls -t ${dir}/* | tail -n ${totalDeleteFiles}`);
         for (let file of deletedFiles.split('\n')) {
             unlink(`${file}`, (err) => {
                 if (err) throw err;
             });
         }
+        return true;
     } catch (e) {
         console.log(e)
     }
